requirements.txt
streamlit>=1.30.0
pydantic>=2.5.0
requests>=2.31.0
numpy>=1.26.0
pandas>=2.1.0
yfinance>=0.2.30
anthropic>=0.40.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
